A slightly modified patch from Praveen Agrawal for "Order Return Process doesn't override the Order Adjustment Amount that is provided from Return Screen GUI" https://issues.apache.org/jira/browse/OFBIZ-5439
OFBiz return services (OrderReturnServices.java - createReturnAdjustment) contains a bug when generating the Return Adjustments based on the Order Adjustments.
The bug is it is ignoring the 'overriding amount' entered on the screen.
For example If an order has an Shipping Charge Order Adjustment in the amount of 7.95 and from the screen the cleint wants to return only 5.95 (override), the return services ignores the 5.95 entry and returns the full 7.95.

jleroux: I removed a BigDecimal amount declaration which was conflicting
I formatted the if block

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1552290 13f79535-47bb-0310-9956-ffa450edef68