A patch from  Deepak Dixit for "If orderDecimalQuantity set to N then system should return error if user add partial quantity in order" https://issues.apache.org/jira/browse/OFBIZ-5962
If productStore.orderDecimalQuantity OR product.orderDecimalQuantity is set to N, then system should return error if we try to add to partial quantity then system should return error instead of rounding it to 0, 
Actual Behavior: If partial quantity is not allowed in order process, and user enter 1.5 quantity then it rounded it and add 2 quantity.
Expected Behavior: if partial quantity is not allowed then system should return error instead of doing rounding if partial quantity passed during add/update item.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1661862 13f79535-47bb-0310-9956-ffa450edef68