Fix regression caused by refactoring start comp - ref OFBIZ-7167
Thanks to an excellent investigation by Jacques Le Roux we realized
that if not all tests are successful (on java -jar ofbiz -jar --tests)
then the system hangs and does not print HTML reports.

After investigating this issue, I realized that the old logic before
refactoring used to call System.exit(99) which executes all the shutdown
hooks and releases the system resources. In the new code, however, the
resources are not released as the exception simply bubbles up to the main
method and thrown out, while the other threads are still running. So
the resources are not released and build.xml cannot continue processing
the HTML results with <junitreport ...>

The fix is simple, but the investigation was hard. Big thank you to Jacques.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1746737 13f79535-47bb-0310-9956-ffa450edef68