remove unit tests from the start component that cause a regression - OFBIZ-7897
There were two tests in the start component that have a hidden regression which
can be repeated as follows:

- start ofbiz with a normal ./gradlew "ofbiz --start"
- Make some changes to the code
- stop ofbiz with ./gradlew "ofbiz --shutdown"

OFBiz will fail to stop because the unit tests will fail due to changed code
while OFBiz is still active.

The correct solution is to convert these tests to integration tests and handle the different
scenarios while the AdminServer is up.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1757360 13f79535-47bb-0310-9956-ffa450edef68